removed redundant class and copied the code over to a more suitable place
