+++ Assets/Scenes/Gameplay Screen.unity
@@ -5759,7 +5759,6 @@ GameObject:
   - component: {fileID: 1551436321}
   - component: {fileID: 1551436320}
   - component: {fileID: 1551436322}
-  - component: {fileID: 1551436323}
   m_Layer: 5
   m_Name: DragPanel
   m_TagString: Untagged
@@ -5845,18 +5844,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   dragPanel: {fileID: 1551436318}
---- !u!114 &1551436323
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1551436318}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 362d1f263170aaa45881e507fdb7555c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
   towerIcons:
   - {fileID: 1169422333}
   - {fileID: 662208631}
