--- Assets/Scripts/TowerIconPositions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 362d1f263170aaa45881e507fdb7555c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
